Reorganize table declarations
