+++ packages/grid/apps/domain/src/main/core/manager/database_manager.py
@@ -1,12 +1,9 @@
 # stdlib
-from typing import Dict
 from typing import List
-from typing import Type
 from typing import Union
 
 # grid relative
 from ..database import BaseModel
-from ..database import db
 
 
 class DatabaseManager:
