+++ packages/grid/apps/domain/src/main/core/manager/request_manager.py
@@ -8,10 +8,9 @@ from syft.core.common.uid import UID
 from syft.core.node.domain.service import RequestStatus
 
 # grid relative
-from ..database.requests.request import Request
+from ..database.tables.request import Request
 from ..exceptions import RequestError
 from .database_manager import DatabaseManager
-from .role_manager import RoleManager
 
 
 class RequestManager(DatabaseManager):
