+++ packages/grid/apps/domain/src/main/core/services/infra_service.py
@@ -1,18 +1,13 @@
 # stdlib
 from datetime import datetime
 import os
-import random
-import secrets
 from typing import List
 from typing import Type
 from typing import Union
 
 # third party
 from nacl.encoding import HexEncoder
-from nacl.signing import SigningKey
 from nacl.signing import VerifyKey
-import requests
-from syft import deserialize
 from syft import serialize
 from syft.core.common.message import ImmediateSyftMessageWithReply
 
@@ -20,11 +15,8 @@ from syft.core.common.message import ImmediateSyftMessageWithReply
 from syft.core.node.abstract.node import AbstractNode
 from syft.core.node.common.service.auth import service_auth
 from syft.core.node.common.service.node_service import ImmediateNodeServiceWithReply
-from syft.core.node.common.service.node_service import ImmediateNodeServiceWithoutReply
-from syft.core.node.domain.client import DomainClient
 from syft.grid.client.client import connect
 from syft.grid.client.grid_connection import GridHTTPConnection
-from syft.grid.connections.http_connection import HTTPConnection
 from syft.grid.messages.infra_messages import CreateWorkerMessage
 from syft.grid.messages.infra_messages import CreateWorkerResponse
 from syft.grid.messages.infra_messages import DeleteWorkerMessage
@@ -35,10 +27,9 @@ from syft.grid.messages.infra_messages import GetWorkerMessage
 from syft.grid.messages.infra_messages import GetWorkerResponse
 from syft.grid.messages.infra_messages import GetWorkersMessage
 from syft.grid.messages.infra_messages import GetWorkersResponse
-from syft.proto.core.io.address_pb2 import Address as Address_PB
 
 # grid relative
-from ...core.database.environment.environment import states
+from ..database.tables.environment import states
 from ...core.infrastructure import AWS_Serverfull
 from ...core.infrastructure import AZURE
 from ...core.infrastructure import Config
