+++ packages/grid/apps/domain/tests/test_routes/test_infrastructure.py
@@ -11,8 +11,8 @@ from src.main.core.database import Role
 from src.main.core.database import Group
 from src.main.core.database import UserGroup
 
-from src.main.core.database.environment.environment import Environment
-from src.main.core.database.environment.user_environment import UserEnvironment
+from src.main.core.database.tables.environment import Environment
+from src.main.core.database.tables.user_environment import UserEnvironment
 
 JSON_DECODE_ERR_MSG = (
     "Expecting property name enclosed in " "double quotes: line 1 column 2 (char 1)"
